Fix: Fix Modal
- change directory
- connect backend
